move test from `simple` to `visitable_group`
